views_mock/requirements.txt
pip==22.3.1
django==4.1.3
djangorestframework==3.14.0
mysqlclient==2.1.1
djangorestframework-simplejwt==5.2.2